requirements.txt
gdown
gymnasium
onnx
onnxruntime
Pillow
streamlit
matplotlib
opencv-python-headless
nvidia-pyindex
matplotlib>=3.3.0
numpy>=1.22.2
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
ray[rllib]
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0
psutil
py-cpuinfo
PySide6
thop>=0.1.1
ultralytics
timm
streamlit
torch
torchvision
pillow
opencv-python
matplotlib
numpy
